(ref): Refactor modal daisy-ui classes into Modal component
