+++ components/modals/ChargeWalletModal.tsx
@@ -1,5 +1,6 @@
 import { AlertError, AlertSuccess } from "components/Alert";
 import Loader from "components/Loader";
+import Modal from "components/Modal";
 import { useSigningClient } from "contexts/cosmwasm";
 import { env } from "env";
 import { useBalance } from "hooks/useBalance";
@@ -80,67 +81,55 @@ export default function ChargeWalletModal({
   }
 
   return (
-    <>
-      <input type="checkbox" id={`charge-modal-${walletAddress}`} className="modal-toggle" />
-      <div className="modal">
-        <div className="modal-box relative">
-          <label
-            htmlFor={`charge-modal-${walletAddress}`}
-            onClick={handleCloseModal}
-            className="btn btn-sm btn-circle absolute right-2 top-2"
-          >
-            ✕
-          </label>
-          <h3 className="text-xl font-bold">Charge your Smart Contract Wallet</h3>
-          <h4 className="text-lg">Available inside your personal wallet: {balance}</h4>
-          <div className="flex flex-col items-center">
-            {!isSending ? (
-              <>
-                <div className="flex flex-col w-full items-start">
-                  <div className="relative rounded-full shadow-sm w-full mt-5">
-                    <input
-                      type="number"
-                      className={`input input-bordered focus:input-primary input-lg w-full pr-24 rounded-full text-center font-mono text-lg ${
-                        getValueValidationError("amountToSend") ? "input-error" : ""
-                      }`}
-                      placeholder="Amount to charge"
-                      step="0.1"
-                      min="0"
-                      onChange={(event) => setAmountToSend(event.target.value)}
-                      value={amountToSend}
-                    />
-                    <span className="absolute top-0 right-0 bottom-0 px-4 py-5 rounded-r-full bg-secondary text-base-100 text-sm">
-                      {convertFromMicroDenom(env.stakingDenom)}
-                    </span>
-                  </div>
-                  {getValueValidationError("amountToSend") && (
-                    <span className="pl-6 text-error font-bold">{getValueValidationError("amountToSend")}</span>
-                  )}
-                </div>
-                <div className="btn btn-primary mt-5 text-xl rounded-full" onClick={chargeWallet}>
-                  Charge
-                </div>
-              </>
-            ) : (
-              <>
-                <Loader>Sending funds to the wallet...</Loader>
-              </>
-            )}
-
-            {sendSuccess && (
-              <div className="mt-4 flex flex-col w-full max-w-xl">
-                <AlertSuccess>{sendSuccess}</AlertSuccess>
+    <Modal id={`charge-modal-${walletAddress}`} onClose={handleCloseModal}>
+      <h3 className="text-xl font-bold">Charge your Smart Contract Wallet</h3>
+      <h4 className="text-lg">Available inside your personal wallet: {balance}</h4>
+      <div className="flex flex-col items-center">
+        {!isSending ? (
+          <>
+            <div className="flex flex-col w-full items-start">
+              <div className="relative rounded-full shadow-sm w-full mt-5">
+                <input
+                  type="number"
+                  className={`input input-bordered focus:input-primary input-lg w-full pr-24 rounded-full text-center font-mono text-lg ${
+                    getValueValidationError("amountToSend") ? "input-error" : ""
+                  }`}
+                  placeholder="Amount to charge"
+                  step="0.1"
+                  min="0"
+                  onChange={(event) => setAmountToSend(event.target.value)}
+                  value={amountToSend}
+                />
+                <span className="absolute top-0 right-0 bottom-0 px-4 py-5 rounded-r-full bg-secondary text-base-100 text-sm">
+                  {convertFromMicroDenom(env.stakingDenom)}
+                </span>
               </div>
-            )}
+              {getValueValidationError("amountToSend") && (
+                <span className="pl-6 text-error font-bold">{getValueValidationError("amountToSend")}</span>
+              )}
+            </div>
+            <div className="btn btn-primary mt-5 text-xl rounded-full" onClick={chargeWallet}>
+              Charge
+            </div>
+          </>
+        ) : (
+          <>
+            <Loader>Sending funds to the wallet...</Loader>
+          </>
+        )}
 
-            {sendError && (
-              <div className="mt-4 flex flex-col w-full max-w-xl">
-                <AlertError>{sendError}</AlertError>
-              </div>
-            )}
+        {sendSuccess && (
+          <div className="mt-4 flex flex-col w-full max-w-xl">
+            <AlertSuccess>{sendSuccess}</AlertSuccess>
           </div>
-        </div>
+        )}
+
+        {sendError && (
+          <div className="mt-4 flex flex-col w-full max-w-xl">
+            <AlertError>{sendError}</AlertError>
+          </div>
+        )}
       </div>
-    </>
+    </Modal>
   );
 }
