+++ components/modals/SendFundsModal.tsx
@@ -1,6 +1,7 @@
 import { AlertError, AlertSuccess } from "components/Alert";
 import { Input } from "components/Input";
 import Loader from "components/Loader";
+import Modal from "components/Modal";
 import TokenAmount from "components/TokenAmount";
 import { useSigningClient } from "contexts/cosmwasm";
 import { WalletInfoWithBalance } from "contexts/vectis";
@@ -87,78 +88,66 @@ export default function SendFundsModal({ walletInfo, walletAddress, onSentFunds,
   }
 
   return (
-    <>
-      <input type="checkbox" id={`send-modal-${walletAddress}`} className="modal-toggle" />
-      <div className="modal">
-        <div className="modal-box relative">
-          <label
-            htmlFor={`send-modal-${walletAddress}`}
-            onClick={handleCloseModal}
-            className="btn btn-sm btn-circle absolute right-2 top-2"
-          >
-            ✕
-          </label>
-          <h3 className="text-xl font-bold">Send funds to address through wallet</h3>
-          <h4 className="text-lg">
-            Available inside wallet: <TokenAmount token={walletInfo?.balance} />
-          </h4>
-          <div className="flex flex-col items-center">
-            {!isSending ? (
-              <>
-                <div className="my-5 w-full">
-                  <Input
-                    placeholder="Receiver address"
-                    onChange={(event) => setReceiverAddress(event.target.value)}
-                    error={getValueValidationError("receiverAddress")}
-                    value={receiverAddress}
-                    autoComplete="off"
-                  />
-                </div>
-                <div className="flex flex-col w-full items-start">
-                  <div className="relative rounded-full shadow-sm w-full">
-                    <input
-                      type="number"
-                      className={`input input-bordered focus:input-primary input-lg w-full pr-24 rounded-full text-center font-mono text-lg ${
-                        getValueValidationError("amountToSend") ? "input-error" : ""
-                      }`}
-                      placeholder="Amount to send"
-                      step="0.1"
-                      min="0"
-                      onChange={(event) => setAmountToSend(event.target.value)}
-                      value={amountToSend}
-                    />
-                    <span className="absolute top-0 right-0 bottom-0 px-4 py-5 rounded-r-full bg-secondary text-base-100 text-sm">
-                      {convertFromMicroDenom(env.stakingDenom)}
-                    </span>
-                  </div>
-                  {getValueValidationError("amountToSend") && (
-                    <span className="pl-6 text-error font-bold">{getValueValidationError("amountToSend")}</span>
-                  )}
-                </div>
-                <div className="btn btn-primary mt-5 text-xl rounded-full" onClick={sendFunds}>
-                  Send
-                </div>
-              </>
-            ) : (
-              <>
-                <Loader>Sending funds to the receiver...</Loader>
-              </>
-            )}
-
-            {sendSuccess && (
-              <div className="mt-4 flex flex-col w-full max-w-xl">
-                <AlertSuccess>{sendSuccess}</AlertSuccess>
+    <Modal id={`send-modal-${walletAddress}`} onClose={handleCloseModal}>
+      <h3 className="text-xl font-bold">Send funds to address through wallet</h3>
+      <h4 className="text-lg">
+        Available inside wallet: <TokenAmount token={walletInfo?.balance} />
+      </h4>
+      <div className="flex flex-col items-center">
+        {!isSending ? (
+          <>
+            <div className="my-5 w-full">
+              <Input
+                placeholder="Receiver address"
+                onChange={(event) => setReceiverAddress(event.target.value)}
+                error={getValueValidationError("receiverAddress")}
+                value={receiverAddress}
+                autoComplete="off"
+              />
+            </div>
+            <div className="flex flex-col w-full items-start">
+              <div className="relative rounded-full shadow-sm w-full">
+                <input
+                  type="number"
+                  className={`input input-bordered focus:input-primary input-lg w-full pr-24 rounded-full text-center font-mono text-lg ${
+                    getValueValidationError("amountToSend") ? "input-error" : ""
+                  }`}
+                  placeholder="Amount to send"
+                  step="0.1"
+                  min="0"
+                  onChange={(event) => setAmountToSend(event.target.value)}
+                  value={amountToSend}
+                />
+                <span className="absolute top-0 right-0 bottom-0 px-4 py-5 rounded-r-full bg-secondary text-base-100 text-sm">
+                  {convertFromMicroDenom(env.stakingDenom)}
+                </span>
               </div>
-            )}
+              {getValueValidationError("amountToSend") && (
+                <span className="pl-6 text-error font-bold">{getValueValidationError("amountToSend")}</span>
+              )}
+            </div>
+            <div className="btn btn-primary mt-5 text-xl rounded-full" onClick={sendFunds}>
+              Send
+            </div>
+          </>
+        ) : (
+          <>
+            <Loader>Sending funds to the receiver...</Loader>
+          </>
+        )}
 
-            {sendError && (
-              <div className="mt-4 flex flex-col w-full max-w-xl">
-                <AlertError>{sendError}</AlertError>
-              </div>
-            )}
+        {sendSuccess && (
+          <div className="mt-4 flex flex-col w-full max-w-xl">
+            <AlertSuccess>{sendSuccess}</AlertSuccess>
           </div>
-        </div>
+        )}
+
+        {sendError && (
+          <div className="mt-4 flex flex-col w-full max-w-xl">
+            <AlertError>{sendError}</AlertError>
+          </div>
+        )}
       </div>
-    </>
+    </Modal>
   );
 }
