+++ components/modals/VoteModal.tsx
@@ -1,6 +1,7 @@
 import { Vote } from "@dao-dao/types/contracts/cw-proposal-single";
 import { AlertError, AlertSuccess } from "components/Alert";
 import Loader from "components/Loader";
+import Modal from "components/Modal";
 import { useSigningClient } from "contexts/cosmwasm";
 import { Proposal } from "contexts/vectis";
 import { useState } from "react";
@@ -37,53 +38,41 @@ export default function VoteModal({ proposal, multisigAddress, onVote, onClose }
   }
 
   return (
-    <>
-      <input type="checkbox" id={`vote-modal-${proposal.id}`} className="modal-toggle" />
-      <div className="modal">
-        <div className="modal-box relative">
-          <label
-            htmlFor={`vote-modal-${proposal.id}`}
-            className="btn btn-sm btn-circle absolute right-2 top-2"
-            onClick={handleCloseModal}
-          >
-            ✕
-          </label>
-          <h3 className="text-2xl font-bold mb-5">Vote Proposal</h3>
-          <div className="flex flex-col items-center text-xl">
-            <p>Title: {proposal.title}</p>
-            <p>Description: {proposal.description}</p>
-            {/* <p>Expires: {proposal.expires.at_time}</p> */}
-            <p>Approval Threshold: {proposal.threshold.absolute_count.weight} positive votes</p>
+    <Modal id={`vote-modal-${proposal.id}`} onClose={handleCloseModal}>
+      <h3 className="text-2xl font-bold mb-5">Vote Proposal</h3>
+      <div className="flex flex-col items-center text-xl">
+        <p>Title: {proposal.title}</p>
+        <p>Description: {proposal.description}</p>
+        {/* <p>Expires: {proposal.expires.at_time}</p> */}
+        <p>Approval Threshold: {proposal.threshold.absolute_count.weight} positive votes</p>
+      </div>
+      {loading && <Loader>Casting your vote...</Loader>}
+      {!loading && !success && (
+        <>
+          <div className="flex items-center space-x-5 justify-center my-5">
+            <button className="btn btn-success" onClick={() => castVote("yes")}>
+              YES
+            </button>
+            <button className="btn btn-error" onClick={() => castVote("no")}>
+              NO
+            </button>
+            <button className="btn" onClick={() => castVote("abstain")}>
+              ABSTAIN
+            </button>
           </div>
-          {loading && <Loader>Casting your vote...</Loader>}
-          {!loading && !success && (
-            <>
-              <div className="flex items-center space-x-5 justify-center my-5">
-                <button className="btn btn-success" onClick={() => castVote("yes")}>
-                  YES
-                </button>
-                <button className="btn btn-error" onClick={() => castVote("no")}>
-                  NO
-                </button>
-                <button className="btn" onClick={() => castVote("abstain")}>
-                  ABSTAIN
-                </button>
-              </div>
 
-              {error && (
-                <div className="my-5">
-                  <AlertError>Error: {error.message}</AlertError>
-                </div>
-              )}
-            </>
-          )}
-          {success && (
+          {error && (
             <div className="my-5">
-              <AlertSuccess>{success}</AlertSuccess>
+              <AlertError>Error: {error.message}</AlertError>
             </div>
           )}
+        </>
+      )}
+      {success && (
+        <div className="my-5">
+          <AlertSuccess>{success}</AlertSuccess>
         </div>
-      </div>
-    </>
+      )}
+    </Modal>
   );
 }
